+++ bricks/feature_brick/__brick__/{{feature_name.snakeCase()}}/widgets/{{feature_name.snakeCase()}}_body.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';{{#isBloc}}
 import '../bloc/bloc.dart';{{/isBloc}}{{#isProvider}}
-import '../provider/provider.dart';{{/isProvider}}
+import '../provider/provider.dart';{{/isProvider}}{{#isRiverpod}}
+import '../provider/provider.dart';{{/isRiverpod}}
 
-class {{feature_name.pascalCase()}}Body extends StatelessWidget {
+class {{feature_name.pascalCase()}}Body {{#isBloc}}extends StatelessWidget{{/isBloc}}{{#isProvider}}extends StatelessWidget{{/isProvider}}{{#isNone}}extends StatelessWidget{{/isNone}}{{#isRiverpod}}extends ConsumerWidget{{/isRiverpod}} {
   const {{feature_name.pascalCase()}}Body({Key? key}) : super(key: key);
 {{#isBloc}}
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context{{#isRiverpod}}, WidgetRef ref{{/isRiverpod}}) {
     return BlocBuilder<{{feature_name.pascalCase()}}Bloc, {{feature_name.pascalCase()}}State>(
       builder: (context, state) {
         return Text(state.customProperty);
@@ -20,7 +21,12 @@ class {{feature_name.pascalCase()}}Body extends StatelessWidget {
         return Text(state.count.toString());
       },
     );
-  } {{/isProvider}} {{#isNone}}
+  } {{/isProvider}} {{#isRiverpod}}
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final count = ref.watch({{feature_name.camelCase()}}Provider);
+    return Text(count.toString());
+  } {{/isRiverpod}} {{#isNone}}
   @override
   Widget build(BuildContext context) {
     return const Text('{{feature_name.pascalCase()}}Page');