+++ bricks/feature_brick/__brick__/{{feature_name.snakeCase()}}/{{#isRiverpod}}provider{{/isRiverpod}}/provider.dart
@@ -1,0 +1,2 @@
+export 'package:flutter_riverpod/flutter_riverpod.dart';
+export '{{feature_name.snakeCase()}}_provider.dart';