+++ bricks/feature_brick/__brick__/{{feature_name.snakeCase()}}/{{#isRiverpod}}provider{{/isRiverpod}}/{{feature_name.snakeCase()}}_provider.dart
@@ -1,0 +1,12 @@
+import 'package:riverpod/riverpod.dart';
+
+final {{feature_name.camelCase()}}Provider = StateNotifierProvider.autoDispose((ref) {
+  return {{feature_name.pascalCase()}}();
+});
+
+class {{feature_name.pascalCase()}} extends StateNotifier<int> {
+  {{feature_name.pascalCase()}}() : super(0);
+
+  void increment() => state++;
+  void decrement() => state--;
+}