+++ bricks/feature_brick/__brick__/{{feature_name.snakeCase()}}/{{feature_name.snakeCase()}}.dart
@@ -1,3 +1,3 @@
-{{#isBloc}}export 'bloc/bloc.dart';{{/isBloc}}
+{{#isBloc}}export 'bloc/bloc.dart';{{/isBloc}}{{#isProvider}}export 'provider/provider.dart';{{/isProvider}}{{#isRiverpod}}export 'provider/provider.dart';{{/isRiverpod}}
 export 'view/{{feature_name.snakeCase()}}_page.dart';
 export 'widgets/widgets.dart';